Fix potential buffer overflow
Changed all sprintf() functions (4 at lines 195, 673, 1036, 1055) to snprintf() functions with the value of 2048 (the original sprintf() writes to a buffer, s_tmpOutBuf, that was initialized with a size of 2048). The snprintf() function will now write a maximum of 2048 bytes of data to the buffer, which prevents writing past the end of the buffer into memory illegally. This adjustment does not change the functionality of the code as any read from the buffer will stop at the end of the buffer, and any characters that exceed the buffer length are discarded and does not cause the program to fail. A ./configure + make build does not fail or throw an error with the changes.